+++ changelog.d/4367.feature
@@ -1,0 +1,1 @@
+Poll Feature - Create Poll Screen
