+++ vector/src/main/AndroidManifest.xml
@@ -339,7 +339,7 @@
         <activity android:name=".features.spaces.manage.SpaceManageActivity" />
         <activity android:name=".features.spaces.people.SpacePeopleActivity" />
         <activity android:name=".features.spaces.leave.SpaceLeaveAdvancedActivity" />
-        <activity android:name=".features.createpoll.CreatePollActivity" />
+        <activity android:name=".features.poll.create.CreatePollActivity" />
 
         <!-- Services -->
