+++ vector/src/main/java/im/vector/app/core/di/FragmentModule.kt
@@ -26,7 +26,7 @@ import dagger.hilt.android.components.ActivityComponent
 import dagger.multibindings.IntoMap
 import im.vector.app.features.attachments.preview.AttachmentsPreviewFragment
 import im.vector.app.features.contactsbook.ContactsBookFragment
-import im.vector.app.features.createpoll.CreatePollFragment
+import im.vector.app.features.poll.create.CreatePollFragment
 import im.vector.app.features.crypto.keysbackup.settings.KeysBackupSettingsFragment
 import im.vector.app.features.crypto.quads.SharedSecuredStorageKeyFragment
 import im.vector.app.features.crypto.quads.SharedSecuredStoragePassphraseFragment
