+++ vector/src/main/java/im/vector/app/core/di/MavericksViewModelModule.kt
@@ -26,7 +26,7 @@ import im.vector.app.features.call.conference.JitsiCallViewModel
 import im.vector.app.features.call.transfer.CallTransferViewModel
 import im.vector.app.features.contactsbook.ContactsBookViewModel
 import im.vector.app.features.createdirect.CreateDirectRoomViewModel
-import im.vector.app.features.createpoll.CreatePollViewModel
+import im.vector.app.features.poll.create.CreatePollViewModel
 import im.vector.app.features.crypto.keysbackup.settings.KeysBackupSettingsViewModel
 import im.vector.app.features.crypto.quads.SharedSecureStorageViewModel
 import im.vector.app.features.crypto.recover.BootstrapSharedViewModel
