+++ vector/src/main/java/im/vector/app/features/navigation/DefaultNavigator.kt
@@ -40,8 +40,9 @@ import im.vector.app.features.call.conference.JitsiCallViewModel
 import im.vector.app.features.call.conference.VectorJitsiActivity
 import im.vector.app.features.call.transfer.CallTransferActivity
 import im.vector.app.features.createdirect.CreateDirectRoomActivity
-import im.vector.app.features.createpoll.CreatePollActivity
-import im.vector.app.features.createpoll.CreatePollArgs
+import im.vector.app.features.poll.create.CreatePollActivity
+import im.vector.app.features.poll.create.CreatePollArgs
+import im.vector.app.features.poll.create.CreatePollViewModel
 import im.vector.app.features.crypto.keysbackup.settings.KeysBackupManageActivity
 import im.vector.app.features.crypto.keysbackup.setup.KeysBackupSetupActivity
 import im.vector.app.features.crypto.recover.BootstrapBottomSheet
@@ -501,7 +502,10 @@ class DefaultNavigator @Inject constructor(
     }
 
     override fun openCreatePoll(context: Context, roomId: String) {
-        val intent = CreatePollActivity.getIntent(context, CreatePollArgs(roomId = roomId))
+        val intent = CreatePollActivity.getIntent(
+                context,
+                CreatePollArgs(roomId = roomId, minOptionsCount = CreatePollViewModel.MIN_OPTIONS_COUNT)
+        )
         context.startActivity(intent)
     }
