+++ vector/src/main/java/im/vector/app/features/poll/create/CreatePollAction.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package im.vector.app.features.createpoll
+package im.vector.app.features.poll.create
 
 import im.vector.app.core.platform.VectorViewModelAction
