+++ vector/src/main/java/im/vector/app/features/poll/create/CreatePollViewEvents.kt
@@ -14,19 +14,12 @@
  * limitations under the License.
  */
 
-package im.vector.app.features.createpoll
+package im.vector.app.features.poll.create
 
-import com.airbnb.mvrx.MavericksState
+import im.vector.app.core.platform.VectorViewEvents
 
-data class CreatePollViewState(
-        val roomId: String,
-        val question: String = "",
-        val options: List<String> = emptyList(),
-        val canCreatePoll: Boolean = false
-) : MavericksState {
-
-    constructor(args: CreatePollArgs) : this(
-            roomId = args.roomId
-    )
+sealed class CreatePollViewEvents : VectorViewEvents {
+    object Success : CreatePollViewEvents()
+    object EmptyQuestionError : CreatePollViewEvents()
+    data class NotEnoughOptionsError(val requiredOptionsCount: Int) : CreatePollViewEvents()
 }
-
