+++ vector/src/main/res/layout/view_attachment_type_selector.xml
@@ -171,6 +171,7 @@
             android:layout_margin="16dp"
             android:baselineAligned="false"
             android:orientation="horizontal"
+            android:visibility="gone"
             android:weightSum="3">
 
             <LinearLayout
